Move console integrity methods to a Console class to preserve Single responsibility principle
