+++ app.rb
@@ -8,8 +8,10 @@ require_relative 'display'
 require_relative 'services/library_user_creator'
 require_relative 'services/library_book_creator'
 require_relative 'services/library_rental_creator'
+require_relative './console.rb'
 
-class App
+class App < Console
+  super()
   include Display
   include UserCreator
   include BookCreator
@@ -20,49 +22,6 @@ class App
     @books = []
     @rentals = []
     @class = Classroom.new('Form 6 Class')
-  end
-
-  def run
-    puts "'Welcome to the Ruby OOP module console API. Dont by your's truly Alick Nyirenda!"
-    loop do
-      menu
-      option = gets.chomp
-      break if option == '7'
-
-      get_num(option)
-    end
-    puts 'Thank you for using our Library!'
-  end
-
-  def menu
-    puts
-    puts 'Please choose an option by entering a number'
-    puts '1 - List all books'
-    puts '2 - List all people'
-    puts '3 - Create a person'
-    puts '4 - Create a book'
-    puts '5 - Create a rental'
-    puts '6 - List all rentals for a given person id'
-    puts '7 - Exit'
-  end
-
-  def get_num(option)
-    case option
-    when '1'
-      list_all_books
-    when '2'
-      list_all_people
-    when '3'
-      create_person
-    when '4'
-      create_book
-    when '5'
-      create_rental
-    when '6'
-      list_rentals_by_person_id
-    else
-      puts 'Please enter a number between 1 and 7'
-    end
   end
 
   def list_all_books
